backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
websockets>=12.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
pyyaml>=6.0.2
numpy>=1.26.4
soundfile>=0.12.1
faster-whisper>=1.0.0
